TestEnvironment Files
TestEnvironment & TestEnvironmentHelper
* moved some parts of TestEnvironment into TestEnvironmentHelper
* greatly reduced amount of code needed to get the TestEnvironment set up
